MEAS.MNT: Move file from csdevice to measurements.
