+++ siriuspy/siriuspy/measurements/__init__.py
@@ -1,2 +1,6 @@
 from .measurements import MeasEnergy
 from .calculations import CalcEnergy, CalcEmmitance, ProcessImage
+
+del calculations, measurements
+
+__all__ = ('measurements', 'calculations', 'csdev')
